survivors based on 'Sex' and based on 'person'
as the visualization says, the majority of survivors are females. then males then kids. and also the nosurvivors males are 4 times more than survivor males. But nonsurvivor of females just half of the survivor females. I think this is because, the priority had been given to females and children during the evacuation